My-Healer
This macro will check if you are poisoned [in this case if you have a potion it will use the potion, otherwise it will cast "cure" on you], if  your HP are less than the MAXIMUM will start a next check, if you have a potion it will use the potion to heal you, else if you don't have the potion and your HP are less then 100 but more than 60, it will cast HEAL on you, else if your HP are less than 60 it will cast GREATER HEAL on you. This process will iterate until your status is restored.