PVP-OFFENSIVE-COMBO
This macro will pre-cast an Explosion, then will activate "stun punch" and give you a message "GO!" [now you must go on your target to stun it!] – If you stun your target the macro will give you a message "SUCCESS!" and starts to cast the first Mindblast, then it will launch an explo potion [with 1-2 sec. of delay] and cast another Mindblast, on the same target. And then another one.

The mana cost of this sequence is 20 [Explo] + 16 [Mindblast] * 3 = 58 mana
The spells will be casted with the right delay.